# Further speed-up reverse depdencies query