--- a/src/models/krate_reverse_dependencies.sql
+++ b/src/models/krate_reverse_dependencies.sql
@@ -1,36 +1,40 @@
--- Apply pagination to the whole thing
-SELECT *, COUNT(*) OVER () as total FROM (
-    -- Multple dependencies can exist, make it distinct
-    SELECT DISTINCT ON (crate_downloads, crate_name)
-    dependencies.*,
-    crates.downloads AS crate_downloads,
-    crates.name AS crate_name
-    FROM dependencies
-    -- We only want the crates whose *max* version is dependent, so we join on a
-    -- subselect that includes the versions with their ordinal position
-    INNER JOIN (
-        SELECT DISTINCT ON (crate_id)
-           *
-        FROM versions
-        WHERE NOT yanked
-        ORDER BY
-            crate_id,
-            semver_no_prerelease DESC NULLS LAST,
-            id DESC
-    ) versions
-      ON versions.id = dependencies.version_id
-    INNER JOIN crates
-      ON crates.id = versions.crate_id
-    WHERE dependencies.crate_id = $1
-    -- this ORDER BY is redundant with the outer one but benefits
-    -- the `DISTINCT ON`
+SELECT
+    dependencies.*, crate_downloads, crate_name, total
+FROM (
+    -- Apply pagination to the crates
+    SELECT *, COUNT(*) OVER () as total FROM (
+        SELECT
+            crates.downloads AS crate_downloads,
+            crates.name AS crate_name,
+            versions.id AS version_id
+        FROM
+        -- We only want the crates whose *max* version is dependent, so we join on a
+        -- subselect that includes the versions with their ordinal position
+        (
+            SELECT DISTINCT ON (crate_id)
+               crate_id, semver_no_prerelease, id
+            FROM versions
+            WHERE NOT yanked
+            ORDER BY
+                crate_id,
+                semver_no_prerelease DESC NULLS LAST,
+                id DESC
+        ) versions
+        INNER JOIN crates
+          ON crates.id = versions.crate_id
+        WHERE versions.id IN (SELECT version_id FROM dependencies WHERE crate_id = $1)
+    ) c
     ORDER BY
         crate_downloads DESC,
-        crate_name ASC,
-        dependencies.id ASC
-) t
-ORDER BY
-    crate_downloads DESC,
-    crate_name ASC
+        crate_name ASC
+) crates
+-- Multiple dependencies can exist, we only want first one
+CROSS JOIN LATERAL (
+    SELECT dependencies.*
+    FROM dependencies
+    WHERE dependencies.crate_id = $1 AND dependencies.version_id = crates.version_id
+    ORDER BY id ASC
+    LIMIT 1
+) dependencies
 OFFSET $2
 LIMIT $3